Fix #23
