+++ src/pydom/context/standard/transformers/inner_html_transformer.py
@@ -1,6 +1,6 @@
 from pydom.rendering.tree.nodes import TextNode
 
-from ...transformers import PropertyTransformer
+from ....rendering.transformers import PropertyTransformer
 
 
 class InnerHTMLTransformer(PropertyTransformer):
