+++ src/pydom/context/standard/transformers/style_transformer.py
@@ -1,5 +1,5 @@
 from ....styling import StyleSheet
-from ...transformers import PropertyTransformer
+from ....rendering.transformers import PropertyTransformer
 
 
 class StyleTransformer(PropertyTransformer):
