+++ src/pydom/rendering/transformers/__init__.py
@@ -1,4 +1,9 @@
-from .post_render_transformer import post_render_transformer
-from .property_transformer import property_transformer
+from .post_render_transformer import post_render_transformer, PostRenderTransformer
+from .property_transformer import property_transformer, PropertyTransformer
 
-__all__ = ["post_render_transformer", "property_transformer"]
+__all__ = [
+    "post_render_transformer",
+    "PostRenderTransformer",
+    "property_transformer",
+    "PropertyTransformer",
+]
