+++ src/pydom/rendering/transformers/post_render_transformer.py
@@ -1,12 +1,27 @@
-from typing import List, Optional, Type, Union
+from abc import ABC, abstractmethod
+from typing import List, Optional, Type, Union, TYPE_CHECKING, Callable
 
-from ...context.context import PostRenderTransformerFunction, get_context, Context
-from ...context.transformers import PostRenderTransformer
+from typing_extensions import ParamSpec, TypeAlias, Concatenate
+
+
+if TYPE_CHECKING:
+    from pydom.context import Context
+    from pydom.rendering.tree.nodes import ContextNode
+
+P = ParamSpec("P")
+
+PostRenderTransformerFunction: TypeAlias = Callable[Concatenate["ContextNode", P], None]
+
+
+class PostRenderTransformer(ABC):
+    @abstractmethod
+    def transform(self, element: "ContextNode"):
+        pass
 
 
 def post_render_transformer(
     *,
-    context: Union[Context, None] = None,
+    context: Union["Context", None] = None,
     before: Optional[List[Type[PostRenderTransformer]]] = None,
     after: Optional[List[Type[PostRenderTransformer]]] = None,
 ):
@@ -30,6 +45,8 @@ def post_render_transformer(
         ...     )
 
     """
+    from ...context import get_context
+
     context = get_context(context)
 
     def decorator(func: PostRenderTransformerFunction):
